fastapi-project/requirements.txt
# Core Framework
fastapi==0.115.0          # Modern web framework - fast, async, auto-generates API docs
uvicorn[standard]==0.32.0 # ASGI server to run FastAPI (like a web server)

# Database
sqlalchemy==2.0.35        # ORM (Object Relational Mapper) - talk to database using Python objects
alembic==1.13.3          # Database migrations - track database changes over time

# Authentication & Security
python-jose[cryptography]==3.3.0  # JWT token creation and verification
passlib[bcrypt]==1.7.4            # Password hashing - never store plain passwords!
python-multipart==0.0.12          # Allows FastAPI to handle form data

# Validation & Settings
pydantic==2.9.2           # Data validation (built into FastAPI, but explicit version)
pydantic-settings==2.6.0  # Load settings from environment variables
python-dotenv==1.0.1      # Read .env files for local development

# Email validation
email-validator==2.2.0    # Validate email addresses in Pydantic models
